perapihan code ulang... still continue...
