+++ cookbook/Database ORM.txt
@@ -1,0 +1,37 @@
+# Save to database
+new_user = User(name=data['name'], email=data['email'], age=data['age'])
+    
+db.session.add(new_user)
+db.session.commit()
+
+# Get all item
+users = User.query.all()
+
+# Get item by id
+user = User.query.get(user_id)
+
+# Update data
+user = User.query.get(user_id)
+
+name = user.name
+email = user.email
+age = user.age
+
+db.session.commit()
+
+# Delete data
+user = User.query.get(user_id)
+
+db.session.delete(user)
+db.session.commit()
+
+# Filter berdasarkan text input
+name = "input_for_filter"
+users = User.query.filter(User.name.like(f"%{name}%")).all()
+
+# Mengurutkan data
+users = User.query.order_by(User.age.desc()).all()
+
+# Pagination
+per_page = 5  # Jumlah user per halaman
+users = User.query.paginate(page=page, per_page=per_page, error_out=False)
