--- src/templates/analytic_templates/index.html
@@ -1,145 +1,0 @@
-{% extends "/layouts/layout.html" %}
-
-{% block content %}
-<div class="container-fluid mt-4">
-    <div class="row" id="layout-container">
-        <!-- Kolom 1: Chatbot -->
-        <div class="col-6 mb-3" id="chat-container" style="display: none;">
-            <div class="card">
-                <div class="card-body">
-                    <div id="chat-header" class="d-flex align-items-center">
-                        <img src="https://iili.io/JpjR5tp.png" alt="Chatbot Image" style="height: 40px;">
-                        <div class="header-text mx-2">
-                            <span>Talita</span>
-                            <span id="typing-indicator"
-                                style="font-size: 12px; color: #888; display: none;">mengetik...</span>
-                        </div>
-                        <button id="minimize-chat" class="btn btn-sm btn-outline-secondary ms-auto">
-                            <i class="fas fa-window-minimize"></i> Minimize
-                        </button>
-                    </div>
-                    <div id="chat-body" class="bg-light p-3" style="height: 400px; overflow-y: auto;">
-                        <!-- Riwayat Pesan Chat akan dimuat di sini -->
-                    </div>
-                    <div id="chat-input" class="d-flex mt-2">
-                        <textarea class="form-control me-2" id="chat-user-input" placeholder="Type your question..."
-                            rows="1"></textarea>
-                        <button id="send-btn" class="btn btn-primary"><i class="fas fa-paper-plane"></i></button>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Kolom 2: Tabel -->
-        <div class="col-12" id="table-container">
-            <div class="card">
-                <div class="card-body">
-                    <h5 class="mb-4">Daftar Perangkat</h5>
-                    <table class="table table-bordered">
-                        <thead>
-                            <tr>
-                                <th>#</th>
-                                <th>Nama Perangkat</th>
-                                <th>Alamat IP</th>
-                                <th>Vendor</th>
-                                <th>Status</th>
-                                <th>Aksi</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <tr>
-                                <td>1</td>
-                                <td>Router 1</td>
-                                <td>192.168.1.1</td>
-                                <td>Cisco</td>
-                                <td>Aktif</td>
-                                <td>
-                                    <button class="btn btn-info btn-sm btn-analytic" data-device-id="1">
-                                        Analytic
-                                    </button>
-                                </td>
-                            </tr>
-                            <tr>
-                                <td>2</td>
-                                <td>Switch 1</td>
-                                <td>192.168.1.2</td>
-                                <td>HP</td>
-                                <td>Nonaktif</td>
-                                <td>
-                                    <button class="btn btn-info btn-sm btn-analytic" data-device-id="2">
-                                        Analytic
-                                    </button>
-                                </td>
-                            </tr>
-                        </tbody>
-                    </table>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
-
-{% block script_js %}
-<!-- JavaScript -->
-<script>
-    document.addEventListener("DOMContentLoaded", function () {
-        const analyticButtons = document.querySelectorAll(".btn-analytic");
-        const chatContainer = document.getElementById("chat-container");
-        const tableContainer = document.getElementById("table-container");
-        const chatBody = document.getElementById("chat-body");
-        const minimizeButton = document.getElementById("minimize-chat");
-
-        // Simpan histori data untuk setiap perangkat
-        const deviceHistories = {
-            1: [
-                "Perangkat Router 1 terdeteksi aktif.",
-                "Analitik terakhir dilakukan pada 03-01-2025 pukul 10:00.",
-                "Tidak ada anomali ditemukan."
-            ],
-            2: [
-                "Perangkat Switch 1 terdeteksi nonaktif.",
-                "Analitik terakhir dilakukan pada 02-01-2025 pukul 14:00.",
-                "Ditemukan masalah konektivitas."
-            ]
-        };
-
-        // Event listener untuk tombol Analytic
-        analyticButtons.forEach(button => {
-            button.addEventListener("click", function () {
-                const deviceId = this.getAttribute("data-device-id");
-
-                // Tampilkan kolom Chatbot
-                chatContainer.style.display = "block";
-                tableContainer.classList.replace("col-12", "col-6");
-
-                // Hapus chat sebelumnya
-                chatBody.innerHTML = "";
-
-                // Muat histori perangkat ke dalam chat
-                if (deviceHistories[deviceId]) {
-                    deviceHistories[deviceId].forEach(message => {
-                        const messageElement = document.createElement("div");
-                        messageElement.classList.add("message", "bot-message", "d-flex", "mb-2");
-                        messageElement.innerHTML = `
-                                <img src="https://iili.io/JpjR5tp.png" alt="Bot Avatar" style="height: 40px; margin-right: 10px;">
-                                <div class="message-text">${message}</div>
-                            `;
-                        chatBody.appendChild(messageElement);
-                    });
-                }
-
-                // Scroll ke bagian bawah chat
-                chatBody.scrollTop = chatBody.scrollHeight;
-            });
-        });
-
-        // Event listener untuk tombol Minimize
-        minimizeButton.addEventListener("click", function () {
-            // Sembunyikan kolom Chatbot
-            chatContainer.style.display = "none";
-            tableContainer.classList.replace("col-6", "col-12");
-        });
-    });
-</script>
-{% endblock %}
