+++ src/templates/device_managers/index.html
@@ -77,7 +77,7 @@
                 <i class="fas fa-trash"></i>
             </button>
         </form>
-        <a href="{{ url_for('ai.analyze_device', device_id=data.id) }}" class="btn btn-warning btn-sm mb-1">AI</a>
+        <a href="{{ url_for('ai_agent_bp.analyze_view', device_id=data.id) }}" class="btn btn-warning btn-sm mb-1">AI</a>
     </td>
 </tr>
