+++ src/templates/layouts/layout_script_func.html
@@ -100,7 +100,7 @@
         typingIndicator.style.display = 'inline'; // Show typing indicator
 
         try {
-            const response = await fetch('/chatbot/chat', {
+            const response = await fetch('/ai/chatbot', {
                 method: 'POST',
                 headers: {
                     'Content-Type': 'application/json'
